+++ src/desktop/commondelegate.cpp
@@ -45,7 +45,7 @@ namespace tremotesf
             progressBar.rect = rect;
             progressBar.minimum = 0;
             progressBar.maximum = 100;
-            progressBar.progress = index.data(mProgressBarRole).toDouble() * 100;
+            progressBar.progress = static_cast<int>(index.data(mProgressBarRole).toDouble() * 100);
             if (progressBar.progress == 0) {
                 progressBar.progress = 1;
             }
